backend/requirements.txt
asgiref==3.8.1
barcode==1.0.4
CacheControl==0.14.3
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-appconf==1.0.6
django-bootstrap-modal-forms==3.0.5
django-environ==0.12.0
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
Levenshtein==0.25.1
msgpack==1.1.0
phonenumberslite==8.13.45
pillow==10.4.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
python-barcode==0.15.1
python-Levenshtein==0.25.1
pytz==2024.2
rapidfuzz==3.9.7
requests==2.32.3
rsa==4.9.1
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.45.1
